requirements.txt
# Polymarket Trading Framework Requirements

# Polymarket Client
py-clob-client>=0.27.0

# Async & WebSocket
websockets>=12.0
aiohttp>=3.9.0
asyncio>=3.4.3

# HTTP Requests
requests>=2.31.0
httpx>=0.25.0

# Environment & Config
python-dotenv>=1.0.0
pyyaml>=6.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0

# Logging & Monitoring
colorlog>=6.7.0

# Utils
python-dateutil>=2.8.2
pytz>=2023.3
